backend/app/requirements.txt
fastapi==0.111.0
uvicorn==0.30.6
pydantic==2.12.3
python-multipart==0.0.20
python-dotenv==1.1.1
google-generativeai==0.8.5
